Warning message about Python version fails in Python3
The `print` command is not executed in Python3, so the PyRate.py module fails
and there is no chance to show the message that PyRate runs in Python2.

This is one of the error messages under python3:

```python
 File "PyRate.py", line 33
      \n""" % (version, build)
              ^
              SyntaxError: invalid syntax
```

I have converted the `print "text", variable` commands to:

```python
print("text {}".format(variable)
```

This method of printing text to screen works under both Python 2 and 3 and has
been implemented since Python2.6.